Fix serde support for DefaultNumericTypes
I was trying to add `evalexpr::HashMapContext<evalexpr::DefaultNumericTypes>`
to one of my serializable structs and noticed that even with serde
enabled, my code wouldn't compile.

I did some poking around in `evalexpr` and concluded that I wouldn't be
able to use `evalexpr` unless I made DefaultNumericTypes serializable.
That's what this patch does. There may be a better way to do this, but I
think this is pretty neat, so I'm sharing it in case it's useful.